requirements.txt
fastapi
uvicorn[standard]
jinja2
python-multipart
python-dotenv
requests
passlib[bcrypt]
python-jose[cryptography]
psycopg2-binary
email-validator
Pillow
apscheduler
pytz
